# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CreateStoryRequest:
    properties:
      city:
        type: string
      country:
        type: string
      preferences:
        items:
          type: string
        type: array
      religions:
        items:
          type: string
        type: array
    type: object
...
  handlers.StoryData:
    properties:
      audio:
        type: string
      audio_type:
        type: string
      image:
        type: string
      story_id:
        type: string
      story_text:
        type: string
      theme:
        type: string
      title:
        type: string
    type: object
  handlers.StoryResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
  model.UserProfile:
    properties:
      city:
        type: string
      country:
        type: string
      email:
        type: string
      preferences:
        items:
          type: string
        type: array
      religions:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  util.HttpError:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for generating and managing stories with AI
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Story API
  version: "1.0"
paths:
...
  /stories:
    get:
      consumes:
      - application/json
      description: Lists stories for a user based on their profile and theme preference.
        Returns stories with signed URLs for images and audio.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Theme filter (1, 2, or 3)
        in: query
        name: theme
        type: string
      - description: 'Number of stories to return (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.StoryData'
            type: array
        "401":
          description: Invalid or missing authorization token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.HttpError'
      summary: List Generated Stories
      tags:
      - Stories
  /story:
    post:
      consumes:
      - application/json
      description: Creates a new story based on the user's profile and preferences.
      parameters:
      - description: Story creation request
        in: body
        name: story
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateStoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Story created successfully
          schema:
            $ref: '#/definitions/handlers.StoryResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/util.HttpError'
        "401":
          description: Invalid or missing authorization token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.HttpError'
      security:
      - BearerAuth: []
      summary: Create a new story
      tags:
      - Story
...
  /user-profile:
    delete:
      consumes:
      - application/json
      description: Deletes the user profile for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: User profile deleted successfully
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Invalid or missing authorization token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.HttpError'
      security:
      - BearerAuth: []
      summary: Delete User Profile
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Gets the user profile information for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Invalid or missing authorization token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.HttpError'
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates the user profile for the authenticated user.
      parameters:
      - description: User profile request
        in: body
        name: userProfile
        required: true
        schema:
          $ref: '#/definitions/model.UserProfile'
      produces:
      - application/json
      responses:
        "200":
          description: User profile updated successfully
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Invalid or missing authorization token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.HttpError'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"